Replaced UI dependent Game class with NewGame class
